requirements.txt
asgiref==3.3.4
Django==3.2
django-rest-auth==0.9.5
django-tinymce==3.3.0
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
